move: 파일 이동
- core역할을 하는 클래스 domain 폴더 안으로 이동
